+++ src/overlays/NaverPolygon.vue
@@ -34,6 +34,14 @@
         map: null,
       };
     },
+    watch: {
+      moreOptions: {
+        deep: true,
+        handler(newValue) {
+          this.setOptions(newValue);
+        }
+      }
+    },
     methods: {
       /**
        * @returns {number}