+++ src/overlays/NaverPolyline.vue
@@ -35,6 +35,14 @@
         map: null,
       };
     },
+    watch: {
+      moreOptions: {
+        deep: true,
+        handler(newValue) {
+          this.setOptions(newValue);
+        }
+      }
+    },
     methods: {
       /**
        * @returns {naver.maps.Bounds}