+++ src/overlays/NaverRectangle.vue
@@ -38,6 +38,14 @@
         map: null,
       };
     },
+    watch: {
+      moreOptions: {
+        deep: true,
+        handler(newValue) {
+          this.setOptions(newValue);
+        }
+      }
+    },
     methods: {
       /**
        * @returns {number}